+++ native/verbx_c/include/verbx_c/cli.h
@@ -1,0 +1,6 @@
+#ifndef VERBX_C_CLI_H
+#define VERBX_C_CLI_H
+
+int verbx_c_run(int argc, char **argv);
+
+#endif